Handle DNT locally
