+++ src/background/deployment.ts
@@ -21,7 +21,8 @@ import { Deployment } from "@/types/contract";
 import browser from "webextension-polyfill";
 import { partition, uniqBy } from "lodash";
 import { reportError } from "@/telemetry/logging";
-import { getExtensionVersion, getUID } from "@/background/telemetry";
+import { getUID } from "@/background/telemetry";
+import { getExtensionVersion } from "@/chrome";
 import { isLinked } from "@/auth/token";
 import { optionsSlice } from "@/options/slices";
 import { reportEvent } from "@/telemetry/events";
