+++ src/hooks/useDeployments.ts
@@ -25,7 +25,8 @@ import { optionsSlice } from "@/options/slices";
 import { selectExtensions } from "@/options/selectors";
 import { getErrorMessage } from "@/errors";
 import useNotifications from "@/hooks/useNotifications";
-import { getExtensionVersion, getUID } from "@/background/telemetry";
+import { getUID } from "@/background/telemetry";
+import { getExtensionVersion } from "@/chrome";
 import { selectInstalledDeployments } from "@/background/deployment";
 import { refreshRegistries } from "@/hooks/useRefresh";
 import { Dispatch } from "redux";
